Run IntelliJ IDEA code formatter

+++ src/main/java/emu/grasscutter/game/managers/ResinManager.java
@@ -1,14 +1,13 @@
 package emu.grasscutter.game.managers;
 
-import static emu.grasscutter.config.Configuration.GAME_OPTIONS;
-
 import emu.grasscutter.game.player.BasePlayerManager;
 import emu.grasscutter.game.player.Player;
 import emu.grasscutter.game.props.PlayerProperty;
 import emu.grasscutter.game.props.WatcherTriggerType;
-import emu.grasscutter.server.packet.send.PacketPlayerPropNotify;
 import emu.grasscutter.server.packet.send.PacketResinChangeNotify;
 import emu.grasscutter.utils.Utils;
+
+import static emu.grasscutter.config.Configuration.GAME_OPTIONS;
 
 public class ResinManager extends BasePlayerManager {
 
@@ -104,7 +103,7 @@ public class ResinManager extends BasePlayerManager {
         // Calculate how much resin we need to refill and update player.
         // Note that this can be more than one in case the player
         // logged off with uncapped resin and is now logging in again.
-        int recharge = 1 + (int)((currentTime - this.player.getNextResinRefresh()) / GAME_OPTIONS.resinOptions.rechargeTime);
+        int recharge = 1 + ((currentTime - this.player.getNextResinRefresh()) / GAME_OPTIONS.resinOptions.rechargeTime);
         int newResin = Math.min(GAME_OPTIONS.resinOptions.cap, currentResin + recharge);
         int resinChange = newResin - currentResin;
 
@@ -114,8 +113,7 @@ public class ResinManager extends BasePlayerManager {
         // Set to zero to disable recharge (because on/over cap.)
         if (newResin >= GAME_OPTIONS.resinOptions.cap) {
             this.player.setNextResinRefresh(0);
-        }
-        else {
+        } else {
             int nextRecharge = this.player.getNextResinRefresh() + resinChange * GAME_OPTIONS.resinOptions.rechargeTime;
             this.player.setNextResinRefresh(nextRecharge);
         }
